backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==1.10.12
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
dnspython==2.6.1
httpx==0.25.0
openai==0.28.1
lxml==5.1.0
bcrypt==4.1.2
pytest==7.4.3
pytest-mock
apify-client>=1.0.0
postmark==1.0.0
